requirements.txt
discord.py
py3-pinterest
